# Optimize FastOr for Bitmaps