This PR really optimizes FastOr on real world data. It does so by

- Pre-calculating and pre-generating a destination Bitmap by analyzing containers in the src bitmaps and their cardinalities.
- Avoiding or reducing memory moves which are expensive. Infact, they are typically the most expensive part of generating bitmaps.
- Avoiding wastage of space by doing a better job at container utilization.
- Expanding the String() method of Bitmap to add more metrics around space amplification, wastage, words/Uid, and memory moves. This analysis is what helped determine the right optimizations above.

Results from this PR vs main:

```
name                                      old time/op    new time/op    delta
RealDataFastOr/census-income_srt-32         12.9ms ± 4%     1.5ms ± 1%  -88.10%  (p=0.008 n=5+5)
RealDataFastOr/census-income-32             13.4ms ± 1%     1.6ms ± 3%  -87.75%  (p=0.008 n=5+5)
RealDataFastOr/census1881_srt-32            18.8ms ± 3%     2.9ms ± 2%  -84.36%  (p=0.008 n=5+5)
RealDataFastOr/census1881-32                25.2ms ± 4%     3.0ms ± 3%  -88.12%  (p=0.008 n=5+5)
RealDataFastOr/dimension_003-32              351ms ± 2%      13ms ± 3%  -96.16%  (p=0.008 n=5+5)
RealDataFastOr/dimension_008-32             96.2ms ± 1%     4.3ms ± 3%  -95.50%  (p=0.008 n=5+5)
RealDataFastOr/dimension_033-32             29.9ms ± 3%     3.5ms ± 3%  -88.19%  (p=0.008 n=5+5)
RealDataFastOr/uscensus2000-32               849µs ± 2%     783µs ± 1%   -7.81%  (p=0.008 n=5+5)
RealDataFastOr/weather_sept_85_srt-32       38.9ms ± 4%     4.2ms ± 2%  -89.26%  (p=0.008 n=5+5)
RealDataFastOr/weather_sept_85-32           45.5ms ± 4%     6.4ms ± 2%  -85.84%  (p=0.008 n=5+5)
RealDataFastOr/wikileaks-noquotes_srt-32    18.9ms ± 3%     1.1ms ± 4%  -94.30%  (p=0.008 n=5+5)
RealDataFastOr/wikileaks-noquotes-32        26.5ms ± 1%     1.1ms ± 2%  -95.78%  (p=0.008 n=5+5)

name                                      old alloc/op   new alloc/op   delta
RealDataFastOr/census-income_srt-32         77.3MB ± 0%     0.1MB ± 0%  -99.91%  (p=0.008 n=5+5)
RealDataFastOr/census-income-32             78.0MB ± 0%     0.1MB ± 0%  -99.91%  (p=0.008 n=5+5)
RealDataFastOr/census1881_srt-32            98.1MB ± 0%     1.5MB ± 0%  -98.49%  (p=0.008 n=5+5)
RealDataFastOr/census1881-32                 151MB ± 0%       1MB ± 0%  -99.02%  (p=0.008 n=5+5)
RealDataFastOr/dimension_003-32             2.80GB ± 0%    0.00GB ± 0%  -99.95%  (p=0.008 n=5+5)
RealDataFastOr/dimension_008-32              724MB ± 0%       0MB ± 0%  -99.98%  (p=0.008 n=5+5)
RealDataFastOr/dimension_033-32              190MB ± 0%       1MB ± 0%  -99.33%  (p=0.008 n=5+5)
RealDataFastOr/uscensus2000-32              1.18MB ± 0%    0.24MB ± 0%  -79.64%  (p=0.008 n=5+5)
RealDataFastOr/weather_sept_85_srt-32        238MB ± 0%       0MB ± 0%  -99.88%  (p=0.008 n=5+5)
RealDataFastOr/weather_sept_85-32            298MB ± 0%       0MB ± 0%  -99.90%  (p=0.008 n=5+5)
RealDataFastOr/wikileaks-noquotes_srt-32     113MB ± 0%       1MB ± 0%  -99.48%  (p=0.008 n=5+5)
RealDataFastOr/wikileaks-noquotes-32         154MB ± 0%       1MB ± 0%  -99.62%  (p=0.008 n=5+5)

name                                      old allocs/op  new allocs/op  delta
RealDataFastOr/census-income_srt-32            728 ± 0%         9 ± 0%  -98.76%  (p=0.008 n=5+5)
RealDataFastOr/census-income-32                799 ± 0%         9 ± 0%  -98.87%  (p=0.008 n=5+5)
RealDataFastOr/census1881_srt-32             2.68k ± 0%     0.03k ± 0%  -98.99%  (p=0.008 n=5+5)
RealDataFastOr/census1881-32                 1.62k ± 0%     0.03k ± 0%  -98.33%  (p=0.008 n=5+5)
RealDataFastOr/dimension_003-32              30.2k ± 0%      0.0k ± 0%  -99.91%  (p=0.000 n=5+4)
RealDataFastOr/dimension_008-32              9.89k ± 0%     0.02k ± 0%  -99.77%  (p=0.008 n=5+5)
RealDataFastOr/dimension_033-32              1.85k ± 0%     0.03k ± 0%  -98.60%  (p=0.008 n=5+5)
RealDataFastOr/uscensus2000-32               1.89k ± 0%     0.06k ± 0%  -97.09%  (p=0.008 n=5+5)
RealDataFastOr/weather_sept_85_srt-32        2.33k ± 0%     0.01k ± 0%  -99.40%  (p=0.008 n=5+5)
RealDataFastOr/weather_sept_85-32            2.89k ± 0%     0.01k ± 0%  -99.52%  (p=0.008 n=5+5)
RealDataFastOr/wikileaks-noquotes_srt-32     1.71k ± 0%     0.02k ± 0%  -99.07%  (p=0.008 n=5+5)
RealDataFastOr/wikileaks-noquotes-32         2.02k ± 0%     0.02k ± 0%  -99.21%  (p=0.008 n=5+5)
```

<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/sroar/4)
<!-- Reviewable:end -->